Version updated 0.86 build 19
We have added a lot of new features to this version of Athanor. However,
we have also solved the max/min issues that plagued the linux version
compiling. The definition of min/max is no longer in atanorconstant.h, and
has been replaced with a local definition in files that need it.