Retrieve atanor cpp version and add it to the versioned libatanor jar name
I did initially consider having the python utility scan the C++ code
in order to retrieve the version, so as not to change the existing
cpp code, however that is highly unreliable and might break with many
legitimate code changes such as enclosing the line in comments, or changing
the varialble names or even spacing in the code. Although it is
possible to write a scanning utility that copes with simple code
changes, the scheme is ultimately flawed. It can only work by
putting documentation warnings for someone not to change their code
in a legitimate way so as not to break a scanner, which is not a
great long term solution.

I implemented a scheme that uses the C++ code to tell us its version
by updating the python version_number.py that we read when we invoke
it during the build which is a much more reliable way of getting
this value . It would have been even better to query the cpp code
via an API rather than invoking a program, but then we get
into python cpp interfaces, and having to build the API before we
can query it, so the implemented scheme is a better compromise than
having to scan the C code to get the version value.