Remove unintended log messages from versioned jar build
I did this last jar build that is renamed into athanor after I
started investigating athanor-server handling of athanor of
log messages. Although I did not check in the modified files,
they were in my local directory and the build still picked them
up.

I can probably add some checks to prevent this in the future,
but it needs the pythongit interface to be installed, and even
then it is difficult to ensure with distributed repositories
that all the commits are pushed etc.