Add some Makefile dependencies
It is tedious and error prone to have to remember to manually build the
athanor java jar libray with an "ant compile" step ater the cpp library
has been built.
Also, there are no include file dependencies, so unless the developer
remembers to do a "make clean", the produced dll or jar may not
reflect the latest status of the code.
I modified the Makefile so that athanor jar file is always built
whenever the cpp dll is built. This takes a few seconds more to
do but it is a safer build.
I also added generated(automated) header file dependencies as manual
dependencies are tedious to maintain. The automated dependencies are
a little difficult to understand and maintain, but it is a better
compromise.
I tested these changes on Linux. I cannot get to run the Makefile
cleanly on Windows 10 even before my changes are incorporated, so
I suspect that may be broken already.
Much more work can be done to unify and tidy up the build and make
it portable (Perhaps by using cmake or another standard tool), but
that's a lot more work, so these are relatively simpler changes that
hopefully improve the build process for the time being.