Too many modifications to take count of.
However, we have modified the current Java library, which now use a new
AtanorCreateGlobal, that create a complete separate space for each loaded
program. Beware, that you can only execute one grammar at a time.

The reason why you cannot run the two grammars in parallel, is due to the
fact that the code reposes on a single global variable which is modified
before the execution. I will see how to modify this behaviour later.